requirements.txt
cryptography==3.4.7
Flask==2.2.2
flask-restful
future==0.18.2
lxml==4.6.3
pandas==1.5.0
pip==21.1.1
py==1.10.0
requests==2.25.1
setuptools==56.0.0
tornado==6.1
wheel==0.37.0
gunicorn
psycopg2-binary